requirements.txt
asgiref==3.9.1
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
jmespath==1.0.1
jwt==1.4.0
numpy==2.3.2
pandas==2.3.2
pgeocode==0.5.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
rest-framework-simplejwt==0.0.2
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
